underhill_attestation: refactoring
